Drop static passwords
For security reasons, all passwords must be generated. That's why
all password related parameters has been moved to defaults but
commented out, so they will be required and one have to set needed
parameters if any of them used but missing, and also to have a reference.

Exclusions:
- `opencontrail_message_queue_password` must be defined due of
  limitations in OpenContrail over OpenStack
- `rabbitmq_guest_password` for backward compatibility
- `keepalived_openstack_telemetry_vip_password` for backward
  compatibility
- `gerrit_ldap_bind_password` for backward compatibility
- `opencontrail_identity_password` for backward compatibility
- `kubernetes_openstack_provider_cloud_password` for backward
  compatibility

This is kind of backport of https://gerrit.mcp.mirantis.com/#/c/34068/
to release/2019.2.0.

Change-Id: Id63bc4be2ef9dfaf369c583e017718c8253c8e93
Prod-related: PROD-29480